corrections and testing
