+++ package.json
@@ -21,12 +21,12 @@
   "dependencies": {
     "@ant-design/icons": "^5.1.4",
     "@ant-design/pro-components": "^2.5.5",
+    "@babel/plugin-transform-private-property-in-object": "^7.23.4",
     "@dfinity/auth-client": "^0.18.1",
     "@electron-toolkit/preload": "^2.0.0",
     "@electron-toolkit/utils": "^1.0.2",
     "@redux-devtools/extension": "^3.2.6",
     "@reduxjs/toolkit": "^1.9.5",
-    "@testing-library/user-event": "^14.5.1",
     "antd": "^5.6.1",
     "antd-style": "^3.3.0",
     "apexcharts": "^3.41.0",
@@ -84,8 +84,9 @@
   "devDependencies": {
     "@babel/plugin-proposal-private-property-in-object": "^7.21.11",
     "@electron/notarize": "^1.2.3",
-    "@testing-library/jest-dom": "^6.1.4",
-    "@testing-library/react": "^14.1.2",
+    "@testing-library/jest-dom": "^5.17.0",
+    "@testing-library/react": "^13.4.0",
+    "@testing-library/user-event": "^14.4.3",
     "@vitejs/plugin-react": "^4.0.0",
     "electron": "^23.0.0",
     "electron-builder": "^23.6.0",
@@ -95,7 +96,8 @@
     "eslint-config-prettier": "^8.8.0",
     "eslint-plugin-prettier": "^4.2.1",
     "eslint-plugin-react": "^7.32.2",
-    "jest": "^29.7.0",
+    "jest": "^29.3.1",
+    "jest-environment-jsdom": "^29.3.1",
     "prettier": "^2.8.8",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
