+++ src/renderer/src/components/dashboard/EventList.jsx
@@ -8,7 +8,7 @@ import { useSelector } from 'react-redux'
 
 const EventList = () => {
   const { customEventListData } = useSelector(eventLogSelector)
-  //console.log(customEventListData)
+  console.log(customEventListData)
 
   return (
     <div style={{ overflow: 'auto', height: 'calc(100vh - 12vh)' }}>
